small fix + reducing warnings
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1517 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Graphic/Common/dwControlledCurve.cpp
@@ -22,7 +22,7 @@ void dwControlledCurve::ComputeCurve()
     if (appliedlimits&&(i==0))
       curve->AddPoint(dwPoint2D((*_controlpoints)[i].GetX(),0));
     curve->AddPoint(dwPoint2D((*_controlpoints)[i].GetX(),(*_controlpoints)[i].GetY()));
-    if (appliedlimits&&(i==_controlpoints->size()-1))
+    if (appliedlimits&&((int)i==_controlpoints->size()-1))
       curve->AddPoint(dwPoint2D((*_controlpoints)[i].GetX(),0));
   }
 /*  if (appliedlimits)
